voters_data/requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
confluent-kafka==2.5.3
idna==3.8
psycopg2-binary==2.9.9
py4j==0.10.9.7
pyspark==3.5.2
requests==2.32.3
simplejson==3.19.3
urllib3==2.2.2
